requirements.txt
asgiref==3.2.10
autopep8==1.5.4
botocore==1.20.97
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coverage==5.2.1
cryptography==3.4.7
decorator==5.0.9
Django==3.1.1
django-admin-honeypot==1.1.0
django-cors-headers==3.5.0
django-defender==0.9.0
django-filter==2.4.0
django-jet==1.0.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.11.1
djangorestframework-simplejwt==4.7.1
dynaconf==3.1.4
hvac==0.10.14
idna==2.10
jmespath==0.10.0
loguru==0.5.3
MarkupSafe==2.0.1
Pillow==8.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
redis==3.5.3
requests==2.25.1
s3transfer==0.4.2
shortuuid==1.0.1
six==1.16.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.5
validators==0.18.2
